RDKB-36072: Hub is not getting IPv6 after upgrade and reboot
Reason for change: Intermittently Hub is failed to configure
IPv6. The reason is dibbler client is not properly restarted
when WanManager set DHCP6.Client DML. Its seems the flag
passed to the thread which used to enable dibbler client has
been changed at runtime when read it inside thread casued the
issue. Correcting this by allocating memory to store the value
and send it to the thread.

Test Procedure:
1. Build should work.
2. IPv4 & IPv6 should work in FR/reboot scenario

Risks: Moderate

(cherry picked from commit b30504eba5f1f3eab59348a3273dd9d842c99300)
(cherry picked from commit aa46bf9717ae457eedf5c11b03fe10cfbc77a066)

+++ source/WanManager/wanmgr_dhcpv6_apis.c
@@ -479,16 +479,24 @@ static int _dibbler_client_operation(char * arg)
 #endif
 
 #ifdef _COSA_BCM_ARM_
-        sprintf(cmd, "killall %s", CLIENT_BIN);
-        system(cmd);
-        sleep(2);
-        sprintf(cmd, "ps -A|grep %s", CLIENT_BIN);
-        _get_shell_output(cmd, out, sizeof(out));
-         if (strstr(out, CLIENT_BIN))
-         {
-            sprintf(cmd, "killall -9 %s", CLIENT_BIN);
+        if (TRUE == WanManager_IsApplicationRunning (CLIENT_BIN))
+        {
+            CcspTraceInfo(("%s-%d [%s] is already running, killing it \n", __FUNCTION__,__LINE__,CLIENT_BIN));
+            sprintf(cmd, "killall %s", CLIENT_BIN);
             system(cmd);
-          }
+            sleep(2);
+#ifdef _HUB4_PRODUCT_REQ_
+            sprintf(cmd, "ps | grep %s | grep -v grep", CLIENT_BIN);
+#else
+            sprintf(cmd, "ps -A|grep %s", CLIENT_BIN);
+#endif // _HUB4_PRODUCT_REQ_
+            _get_shell_output(cmd, out, sizeof(out));
+            if (strstr(out, CLIENT_BIN))
+            {
+                sprintf(cmd, "killall -9 %s", CLIENT_BIN);
+                system(cmd);
+            }
+        }
 #endif
     }
     else if (!strncmp(arg, "start", 5))
